+++ .gitignore
@@ -27,4 +27,4 @@ text-ui-test/data/
 
 data/savedTasks.txt
 Event flow.txt
-src/main/java/exceptions/All wrong cmds.txt
+src/main/java/wessy.exceptions/All wrong cmds.txt
