+++ src/main/java/wessy/CmdType.java
@@ -1,3 +1,5 @@
+package wessy;
+
 import java.util.Map;
 import java.util.HashMap;
 import java.util.Set;
