+++ src/main/java/wessy/task/Deadline.java
@@ -1,4 +1,4 @@
-package task;
+package wessy.task;
 
 import java.time.LocalDateTime;
