+++ src/main/java/wessy/task/ToDo.java
@@ -1,4 +1,4 @@
-package task;
+package wessy.task;
 
 public class ToDo extends Task {
     public ToDo(String description, boolean isDone) {
